* develop: (63 commits)
  Bumped version to 3.8.1
  Bump struts version to 2.5.26
  More HR properties parsing tests
  Fix hr.properties parsing
  Bumped version to 3.8.1-SNAPSHOT
  Fix typo in method name
  Fixed license header
  Bump junit from 4.11 to 4.13.1
  Bumped version to 3.8.0
  Better handling of some RDNs
  Bumped version to 3.8.0.rc0
  Add support for mysql db url parameters
  Fix inconsistent remove certificate behaviour
  Fixed SOAP delete attribute class
  Upgraded to struts 2.5.22
  Fix mysql driver name
  Bump versions for guava, commons-collections and mysql-connector
  Align with EPEL location of voms mysql plugin
  More work towards 3.8.0
  Fixed inconsistency in pom.xml
  ...

+++ voms-admin-server/src/main/java/org/glite/security/voms/admin/persistence/dao/generic/AuditSearchDAO.java
@@ -27,13 +27,13 @@ public interface AuditSearchDAO extends GenericDAO<AuditEvent, Long> {
 
   List<AuditEvent> findLatestEvents(int numEvents);
 
-  Integer countEvents();
+  Long countEvents();
 
   AuditLogSearchResults findEventsMatchingParams(AuditLogSearchParams sp);
 
   ScrollableAuditLogSearchResults scrollEventsMatchingParams(
     AuditLogSearchParams sp, ScrollMode scrollMode);
 
-  Integer countEventsMatchingParams(AuditLogSearchParams sp);
+  Long countEventsMatchingParams(AuditLogSearchParams sp);
 
 }
